# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for Telisik Tryout",
        "title": "Tryout Sample",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Artamananda",
            "email": "artamananda@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.28"
    },
    "host": "localhost:8080",
    "basePath": "/v1/api",
    "paths": {
...
        "/transaction-tryout": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Retrieve a list of all transaction tryouts, optionally filtered by tryoutId and userId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction Tryouts"
                ],
                "summary": "Find all transaction tryouts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tryout ID",
                        "name": "tryoutId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GeneralResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create a new transaction tryout with provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction Tryouts"
                ],
                "summary": "Create a transaction tryout",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateTransactionTryoutRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.GeneralResponse"
                        }
                    }
                }
            }
        },
        "/transaction-tryout/complete": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update the status of a transaction tryout to COMPLETED upon successful validation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction Tryouts"
                ],
                "summary": "Update transaction tryout status to COMPLETED",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateStatusTransactionTryoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GeneralResponse"
                        }
                    }
                }
            }
        },
        "/transaction-tryout/paid": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update the status of a transaction tryout to PAID upon successful token validation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction Tryouts"
                ],
                "summary": "Update transaction tryout status to PAID",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateTransactionTryoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GeneralResponse"
                        }
                    }
                }
            }
        },
        "/transaction-tryout/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Retrieve a transaction tryout by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction Tryouts"
                ],
                "summary": "Find a transaction tryout by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction Tryout ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GeneralResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Delete a transaction tryout by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction Tryouts"
                ],
                "summary": "Delete a transaction tryout by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction Tryout ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GeneralResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateTransactionTryoutRequest": {
            "type": "object",
            "properties": {
                "tryout_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "model.GeneralResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "payload": {}
            }
        },
...
        "model.UpdateStatusTransactionTryoutRequest": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "tryout_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.UpdateTransactionTryoutRequest": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "transaction_tryout_id": {
                    "type": "string"
                },
                "tryout_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}